refactor(tests): refactorizar módulo Users completo - dividir archivos >300 líneas
✨ Refactorización completa del módulo Users

Archivos refactorizados:
- users.service.spec.ts (610 líneas) → 3 archivos especializados
  - users.service.crud.spec.ts (152 líneas): create, findByEmail, findById, findOne
  - users.service.updates.spec.ts (251 líneas): update, updateLastLogin, markEmailAsVerified, remove, activate
  - users.service.query.spec.ts (136 líneas): findAll con filtros y paginación

- users.controller.spec.ts (433 líneas) → 2 archivos especializados
  - users.controller.crud.spec.ts (260 líneas): create, findAll, getProfile, findOne
  - users.controller.updates.spec.ts (193 líneas): update, remove, activate

- Creación de helpers/users.test-helpers.ts (107 líneas):
  - createMockUser(): factory para User entity
  - createMockCreateUserDto(): factory para CreateUserDto
  - createMockUserRepository(): factory para mock Repository
  - createMockQueryBuilder(): factory para QueryBuilder con paginación
  - setupUsersTestModule(): setup completo del módulo de pruebas

Mejoras aplicadas:
- test.each() para reducir duplicación en findAll filters (3 tests → 1 parametrizado)
- Factory pattern para objetos mock complejos (User, DTOs, QueryBuilder)
- Separación por responsabilidad: CRUD, Updates, Queries
- DRY: helpers compartidos entre service y controller tests
- Paginación mock simplificada con createMockQueryBuilder()

Archivos que NO requieren refactor (<300 líneas):
- user.entity.spec.ts (88 líneas) - OK

Resumen del módulo Users:
- Total refactorizado: 2 archivos (1,043 líneas) → 5 archivos especializados
- Promedio: 198 líneas/archivo (target: 150-250)
- Helpers: 107 líneas (factories y utilidades reutilizables)
- Tests: 61 passing (100% pass rate)
  - Service: 52 tests (CRUD: 11, Updates: 16, Queries: 7)
  - Controller: 18 tests (CRUD & Query: 11, Updates: 7)
  - Entity: 9 tests
- Coverage módulo users: 92.15% statements, 100% functions
- Lint: ✅ Clean (no errors)
- TypeScript: ✅ All checks passed

Progreso total Tarea 17:
- Módulo Auth: ✅ COMPLETADO (3 archivos refactorizados)
- Módulo Users: ✅ COMPLETADO (2 archivos refactorizados)
- Próximo: Módulo Products (1 archivo crítico: 1,088 líneas)
