+++ src/modules/users/users.controller.crud.spec.ts
@@ -2,10 +2,10 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
 import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
-import { CreateUserDto, UpdateUserDto, UserQueryDto } from './dto';
+import { CreateUserDto, UserQueryDto } from './dto';
 import { User } from './entities/user.entity';
 
-describe('UsersController', () => {
+describe('UsersController - CRUD & Query Operations', () => {
   let controller: UsersController;
   let service: jest.Mocked<UsersService>;
 
@@ -149,48 +149,12 @@ describe('UsersController', () => {
       expect(service.findAll).toHaveBeenCalledWith(queryDto);
     });
 
-    it('should filter users by search term', async () => {
+    it.each([
+      ['search', { page: 1, limit: 10, search: 'john' }],
+      ['status', { page: 1, limit: 10, status: 'active' as const }],
+      ['sorting', { page: 1, limit: 10, sortBy: 'firstName', sortOrder: 'ASC' as const }],
+    ])('should apply %s filter correctly', async (_, queryDto) => {
       // Arrange
-      const queryDto: UserQueryDto = {
-        page: 1,
-        limit: 10,
-        search: 'john',
-      };
-      service.findAll.mockResolvedValue(mockPaginatedResponse);
-
-      // Act
-      const result = await controller.findAll(queryDto);
-
-      // Assert
-      expect(result).toBeDefined();
-      expect(service.findAll).toHaveBeenCalledWith(queryDto);
-    });
-
-    it('should filter users by status', async () => {
-      // Arrange
-      const queryDto: UserQueryDto = {
-        page: 1,
-        limit: 10,
-        status: 'active',
-      };
-      service.findAll.mockResolvedValue(mockPaginatedResponse);
-
-      // Act
-      const result = await controller.findAll(queryDto);
-
-      // Assert
-      expect(result).toBeDefined();
-      expect(service.findAll).toHaveBeenCalledWith(queryDto);
-    });
-
-    it('should apply custom sorting', async () => {
-      // Arrange
-      const queryDto: UserQueryDto = {
-        page: 1,
-        limit: 10,
-        sortBy: 'firstName',
-        sortOrder: 'ASC',
-      };
       service.findAll.mockResolvedValue(mockPaginatedResponse);
 
       // Act
@@ -291,143 +255,6 @@ describe('UsersController', () => {
       // Assert
       expect(result).toBeDefined();
       expect(service.findOne).toHaveBeenCalledWith(validUuid);
-    });
-  });
-
-  describe('update', () => {
-    it('should update a user successfully', async () => {
-      // Arrange
-      const userId = 'user-123';
-      const updateDto: UpdateUserDto = {
-        firstName: 'UpdatedJohn',
-        lastName: 'UpdatedDoe',
-      };
-      const updatedUser = {
-        ...mockUserResponse,
-        firstName: 'UpdatedJohn',
-        lastName: 'UpdatedDoe',
-        fullName: 'UpdatedJohn UpdatedDoe',
-        updatedAt: new Date(),
-      };
-      service.update.mockResolvedValue(updatedUser);
-
-      // Act
-      const result = await controller.update(userId, updateDto);
-
-      // Assert
-      expect(result).toEqual(updatedUser);
-      expect(result.firstName).toBe('UpdatedJohn');
-      expect(result.lastName).toBe('UpdatedDoe');
-      expect(service.update).toHaveBeenCalledWith(userId, updateDto);
-    });
-
-    it('should update multiple user fields', async () => {
-      // Arrange
-      const userId = 'user-123';
-      const updateDto: UpdateUserDto = {
-        firstName: 'Multi',
-        lastName: 'Update',
-        phoneNumber: '+1234567890',
-      };
-      const updatedUser = {
-        ...mockUserResponse,
-        firstName: 'Multi',
-        lastName: 'Update',
-        fullName: 'Multi Update',
-        phoneNumber: '+1234567890',
-      };
-      service.update.mockResolvedValue(updatedUser);
-
-      // Act
-      const result = await controller.update(userId, updateDto);
-
-      // Assert
-      expect(result.firstName).toBe('Multi');
-      expect(result.lastName).toBe('Update');
-      expect(result.phoneNumber).toBe('+1234567890');
-      expect(service.update).toHaveBeenCalledWith(userId, updateDto);
-    });
-
-    it('should handle partial updates', async () => {
-      // Arrange
-      const userId = 'user-123';
-      const updateDto: UpdateUserDto = {
-        firstName: 'OnlyFirstName',
-      };
-      const updatedUser = {
-        ...mockUserResponse,
-        firstName: 'OnlyFirstName',
-        fullName: 'OnlyFirstName Doe',
-      };
-      service.update.mockResolvedValue(updatedUser);
-
-      // Act
-      const result = await controller.update(userId, updateDto);
-
-      // Assert
-      expect(result.firstName).toBe('OnlyFirstName');
-      expect(result.lastName).toBe(mockUserResponse.lastName);
-      expect(service.update).toHaveBeenCalledWith(userId, updateDto);
-    });
-  });
-
-  describe('remove', () => {
-    it('should soft delete a user', async () => {
-      // Arrange
-      const userId = 'user-123';
-      service.remove.mockResolvedValue(undefined);
-
-      // Act
-      const result = await controller.remove(userId);
-
-      // Assert
-      expect(result).toBeUndefined();
-      expect(service.remove).toHaveBeenCalledWith(userId);
-    });
-
-    it('should call remove with correct UUID', async () => {
-      // Arrange
-      const userId = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11';
-      service.remove.mockResolvedValue(undefined);
-
-      // Act
-      await controller.remove(userId);
-
-      // Assert
-      expect(service.remove).toHaveBeenCalledWith(userId);
-    });
-  });
-
-  describe('activate', () => {
-    it('should activate a deactivated user', async () => {
-      // Arrange
-      const userId = 'user-123';
-      const activatedUser = {
-        ...mockUserResponse,
-        isActive: true,
-      };
-      service.activate.mockResolvedValue(activatedUser);
-
-      // Act
-      const result = await controller.activate(userId);
-
-      // Assert
-      expect(result).toEqual(activatedUser);
-      expect(result.isActive).toBe(true);
-      expect(service.activate).toHaveBeenCalledWith(userId);
-    });
-
-    it('should call activate with correct UUID', async () => {
-      // Arrange
-      const userId = 'f47ac10b-58cc-4372-a567-0e02b2c3d479';
-      service.activate.mockResolvedValue(mockUserResponse);
-
-      // Act
-      const result = await controller.activate(userId);
-
-      // Assert
-      expect(result).toBeDefined();
-      expect(service.activate).toHaveBeenCalledWith(userId);
     });
   });
 });
